+++ test/test_exam_text_case_asserts.py
@@ -54,6 +54,100 @@ class Test_ExamTestCase(unittest.TestCase):
         self.assertEqual(test.tags, [])
 
 
+
+    def test_assert_module_standard_import(self):
+        """
+        Test that assertModule works for checking import of standard library
+        """
+        class TestAssertModule(ExamTestCase):
+            def test_foo(self_):
+                self_.assertModule('random')
+
+        test = TestAssertModule('test_foo')
+        test.test_foo()
+        self.assertEqual(test.correct_answer, "'random'")
+        self.assertEqual(test.student_answer, "None")
+
+
+    def test_assert_module_standard_import_fail(self):
+        """
+        Test that assertModule raise exception for checking import of standard library
+        """
+        class TestAssertModule(ExamTestCase):
+            def test_foo(self_):
+                self_.assertModule('random_not_existing')
+
+        test = TestAssertModule('test_foo')
+        with self.assertRaises(AssertionError):
+            test.test_foo()
+        self.assertEqual(test.correct_answer, "'random_not_existing'")
+        self.assertEqual(test.student_answer, "None")
+
+
+
+    def test_assert_module_from_path(self):
+        """
+        Test that assertModule works for checking import from path
+        """
+        dir_path = os.path.dirname(os.path.realpath(__file__))
+        class TestAssertModule(ExamTestCase):
+            def test_foo(self_):
+                self_.assertModule('__init__', dir_path)
+
+        test = TestAssertModule('test_foo')
+        test.test_foo()
+        self.assertEqual(test.correct_answer, "'__init__'")
+        self.assertEqual(test.student_answer, repr(dir_path))
+
+
+
+    def test_assert_module_from_path_fail(self):
+        """
+        Test that assertModule raise exception for checking import from path
+        """
+        dir_path = os.path.dirname(os.path.realpath(__file__))+"/crazy_path"
+        class TestAssertModule(ExamTestCase):
+            def test_foo(self_):
+                self_.assertModule('__init__', dir_path)
+
+        test = TestAssertModule('test_foo')
+        with self.assertRaises(AssertionError):
+            test.test_foo()
+        self.assertEqual(test.correct_answer, "'__init__'")
+        self.assertEqual(test.student_answer, repr(dir_path))
+
+
+
+    def test_assert_attribute(self):
+        """
+        Test that assertAttribute finds function in module
+        """
+        class TestAssertAttribute(ExamTestCase):
+            def test_foo(self_):
+                self_.assertAttribute(ExamTestCase, "assertAttribute")
+
+        test = TestAssertAttribute('test_foo')
+        test.test_foo()
+        self.assertEqual(test.correct_answer, "'assertAttribute'")
+        self.assertEqual(test.student_answer, "<class 'exam_test_case.ExamTestCase'>")
+
+
+
+    def test_assert_attribute_fail(self):
+        """
+        Test that assertAttribute raise error when moidule is missing function
+        """
+        class TestAssertAttribute(ExamTestCase):
+            def test_foo(self_):
+                self_.assertAttribute(ExamTestCase, "NotAFunction")
+
+        test = TestAssertAttribute('test_foo')
+        with self.assertRaises(AssertionError):
+            test.test_foo()
+        self.assertEqual(test.correct_answer, "'NotAFunction'")
+        self.assertEqual(test.student_answer, "<class 'exam_test_case.ExamTestCase'>")
+
+
 if __name__ == '__main__':
     # runner = unittest.TextTestRunner(resultclass=ExamTestResult, verbosity=2)
     unittest.main(verbosity=2)